+++ src/components/Tooltip/index.ts
@@ -1,0 +1,5 @@
+export { Tooltip } from "@/components/Tooltip/Tooltip";
+export type {
+  TooltipPosition,
+  TooltipProps,
+} from "@/components/Tooltip/interfaces";