+++ src/components/Tooltip/interfaces.ts
@@ -1,0 +1,9 @@
+import type { PropsWithChildren } from "react";
+
+export type TooltipPosition = "above" | "below" | "left" | "right";
+
+export type TooltipProps = PropsWithChildren<{
+  position: TooltipPosition;
+  text: string;
+  disabled: boolean;
+}>;